Move component/componentattribute to internal/telemetry/componentattribute
