+++ cmd/otelcorecol/builder-config.yaml
@@ -41,7 +41,6 @@ replaces:
   - go.opentelemetry.io/collector/client => ../../client
   - go.opentelemetry.io/collector/component => ../../component
   - go.opentelemetry.io/collector/component/componenttest => ../../component/componenttest
-  - go.opentelemetry.io/collector/component/componentattribute => ../../component/componentattribute
   - go.opentelemetry.io/collector/component/componentstatus => ../../component/componentstatus
   - go.opentelemetry.io/collector/config/configauth => ../../config/configauth
   - go.opentelemetry.io/collector/config/configcompression => ../../config/configcompression
