--- component/componentattribute/Makefile
@@ -1,1 +1,0 @@
-include ../../Makefile.Common
