--- component/componentattribute/go.mod
@@ -1,22 +1,0 @@
-module go.opentelemetry.io/collector/component/componentattribute
-
-go 1.22.0
-
-require (
-	github.com/stretchr/testify v1.10.0
-	go.opentelemetry.io/collector/pipeline v0.119.0
-	go.opentelemetry.io/otel v1.34.0
-	go.uber.org/zap v1.27.0
-)
-
-require (
-	github.com/davecgh/go-spew v1.1.1 // indirect
-	github.com/kr/pretty v0.3.1 // indirect
-	github.com/pmezard/go-difflib v1.0.0 // indirect
-	github.com/rogpeppe/go-internal v1.10.0 // indirect
-	go.uber.org/multierr v1.11.0 // indirect
-	gopkg.in/check.v1 v1.0.0-20201130134442-10cb98267c6c // indirect
-	gopkg.in/yaml.v3 v3.0.1 // indirect
-)
-
-replace go.opentelemetry.io/collector/pipeline => ../../pipeline
