+++ internal/telemetry/go.mod
@@ -3,14 +3,18 @@ module go.opentelemetry.io/collector/internal/telemetry
 go 1.22.0
 
 require (
+	github.com/stretchr/testify v1.10.0
 	go.opentelemetry.io/collector/component v0.119.0
+	go.opentelemetry.io/collector/pipeline v0.119.0
+	go.opentelemetry.io/otel v1.34.0
 	go.uber.org/zap v1.27.0
 )
 
 require (
+	github.com/davecgh/go-spew v1.1.1 // indirect
 	github.com/gogo/protobuf v1.3.2 // indirect
+	github.com/pmezard/go-difflib v1.0.0 // indirect
 	go.opentelemetry.io/collector/pdata v1.25.0 // indirect
-	go.opentelemetry.io/otel v1.34.0 // indirect
 	go.opentelemetry.io/otel/metric v1.34.0 // indirect
 	go.opentelemetry.io/otel/trace v1.34.0 // indirect
 	go.uber.org/multierr v1.11.0 // indirect
@@ -20,8 +24,11 @@ require (
 	google.golang.org/genproto/googleapis/rpc v0.0.0-20241202173237-19429a94021a // indirect
 	google.golang.org/grpc v1.70.0 // indirect
 	google.golang.org/protobuf v1.36.5 // indirect
+	gopkg.in/yaml.v3 v3.0.1 // indirect
 )
 
 replace go.opentelemetry.io/collector/pdata => ../../pdata
 
 replace go.opentelemetry.io/collector/component => ../../component
+
+replace go.opentelemetry.io/collector/pipeline => ../../pipeline
