+++ service/go.mod
@@ -10,7 +10,6 @@ require (
 	github.com/stretchr/testify v1.10.0
 	go.opentelemetry.io/collector v0.119.0
 	go.opentelemetry.io/collector/component v0.119.0
-	go.opentelemetry.io/collector/component/componentattribute v0.0.0-20250207221750-83d93cd7cf86
 	go.opentelemetry.io/collector/component/componentstatus v0.119.0
 	go.opentelemetry.io/collector/component/componenttest v0.119.0
 	go.opentelemetry.io/collector/config/confighttp v0.119.0
@@ -32,6 +31,7 @@ require (
 	go.opentelemetry.io/collector/extension/zpagesextension v0.119.0
 	go.opentelemetry.io/collector/featuregate v1.25.0
 	go.opentelemetry.io/collector/internal/fanoutconsumer v0.119.0
+	go.opentelemetry.io/collector/internal/telemetry v0.0.0-00010101000000-000000000000
 	go.opentelemetry.io/collector/pdata v1.25.0
 	go.opentelemetry.io/collector/pdata/pprofile v0.119.0
 	go.opentelemetry.io/collector/pdata/testdata v0.119.0
@@ -138,7 +138,7 @@ replace go.opentelemetry.io/collector/component => ../component
 
 replace go.opentelemetry.io/collector/component/componenttest => ../component/componenttest
 
-replace go.opentelemetry.io/collector/component/componentattribute => ../component/componentattribute
+replace go.opentelemetry.io/collector/internal/telemetry => ../internal/telemetry/
 
 replace go.opentelemetry.io/collector/component/componentstatus => ../component/componentstatus
