+++ versions.yaml
@@ -31,7 +31,7 @@ module-sets:
       - go.opentelemetry.io/collector/cmd/builder
       - go.opentelemetry.io/collector/cmd/mdatagen
       - go.opentelemetry.io/collector/component
-      - go.opentelemetry.io/collector/component/componentattribute
+      - go.opentelemetry.io/collector/internal/telemetry/componentattribute
       - go.opentelemetry.io/collector/component/componentstatus
       - go.opentelemetry.io/collector/component/componenttest
       - go.opentelemetry.io/collector/confmap/xconfmap
